pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anticorrupt"
version = "0.1.0"
description = "AI-Assisted Political & Institutional Explainer Platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # Data & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "networkx>=3.2.1",
    "sqlite-utils>=3.35.2",
    # CLI & UI
    "typer>=0.9.0",
    "rich>=13.7.0",
    "questionary>=2.0.1",
    # AI / LLM
    "anthropic>=0.20.0",
    "openai>=1.10.0",
    "jinja2>=3.1.3",
    # HTTP & Ingestion
    "httpx>=0.26.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "trafilatura>=1.8.0",
    # Visuals
    "pillow>=10.2.0",
    # Utils
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
anticorrupt = "src.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
